+++ desktop/src/com/mygdx/game/AIControl/IMoveable.java
@@ -3,9 +3,9 @@ package com.mygdx.game.AIControl;
 import com.mygdx.game.Entities.AI;
 
 public interface IMoveable {
-    public void moveDown(AI entityAI, int speed, int pos);
-    public void moveUp(AI entityAI, int speed, int pos);
-    public void moveLeft(AI entityAI, int speed, int pos);
-    public void moveRight(AI entityAI, int speed, int pos);
+    public void moveDown(AI entityAI, float speed, float pos);
+    public void moveUp(AI entityAI, float speed, float pos);
+    public void moveLeft(AI entityAI, float speed, float pos);
+    public void moveRight(AI entityAI, float speed, float pos);
     
 }
